Download packages from rsync mirror if local repository is empty.
This is for package lists in "pkglist/" subdirectory where a 3rd party
repository is the package source.
If the pkglists/*.conf file defines a rsync mirror URL in
variable SL_REPO_URL, then instead of aborting with an error when
the local repository directory (SL_REPO) is empty or non-existing,
the packages will be downloaded from that mirror before continuing.