Add GUI partitioning tool (gparted)
The scripts feature an installer (setup2hd) that doesn't
have a partitioning tool.
While users can use cfdisk and install afterwards,
having a GUI tool is more userfriendly.

gparted is only running with root rights.
Since it's not trivial to run it as root
and keep your X session on Slackware, we'll just use bashrun2 for that.