Fix the 'setup2hd' script's method of Live content installation.
In the original version of the script, the squashfs modules' contents
would be extracted one after the other.
However this disregards any package or file deletions inbetween modules.
As a result, e.g. a PLASMA5 installation to disk using 'setup2hd' would
leave several artefacts in removed_scripts and removed_packages and probably
in other locations too.
The overlay is now constructed somewhat differently, so that 'setup2hd' can
access the full read-only filesystem properly; it will copy this content
to the harddisk using rsync.  A progress indicator will show the stats of
files that are actually being transferred.

At the same time, the 'setup2hd' script's language settings are now
hard-coded by setting 'LC_ALL' and LANG to "C".  Slackware's setup scripts
parse command output and expect english texts there, but when you selected
a different language when booting your Live OS, this would confuse e.g.
'liloconfig' to the extent that it failed to find the Linux partitions.