XFCE: add 'squashfs-tools' to the minimal package list (min.lst).
The squashfs-tools are an essential package when using 'setup2hd',
the Slackware Live OS' harddisk installer.
The XFCE ISO does not contain setup2hd at all (it only gets added
to ISOs with a huge kernel) but if someone creates a custom Live OS
containing a huge kernel but based on the minimal XFCE variant,
this custom Live OS requires the presence of unsquashfs,
or else 'setup2hd' will fail.