Use XZ compression for the initrd instead of GZIP.
This shrinks the initrd back from 22 to 17 MB in size.
Also, a redundant copy of the initrd was removed from the
'0000-slackware_boot' module, reducing the ISO size with an
additional 17 MB.

These two modifications compensate for the increased size due to
addition of network modules and firmware (which are needed for
network boot).

Goal reached: the XFCE ISO still fits on CDROM media.