Copy the complete system to disk
We use AlienBob's hook to install /home and more config files,
so the user will have a graphical system instead of a text login
with all application presets preserved.

This also adds a hint about gparted (since we included it)
and fixes up the Xzibit phrasing left in the scripts. (e.g. Cinelive Live)