XFCE: add some network tools, remove some entertainment (to stay below 703MB).
Added packages:
+ bind
+ dnsmasq
+ idnkit
+ traceroute

Removed packages:
- xfractint
- xlockmore
- xmms

The XFCE ISO has xaos as alternative to xfractint. The xlockmore program
is not used anyway (we have xscreensaver).
XMMS was a hard call but there's audacious still.
In return, you get useful stuff like nslookup and host, and dnsmasq is
required by the PXE server (soon to be added).