makemod: add interactive mode (-i switch) to prevent overwriting module file.
Default behaviour of makemod is to overwrite the destination squashfs module
without asking.  When adding the '-i' switch to the 'makemod' command,
the script will ask explicitly whether it is allowed to overwrite an
existing file.
Also added is a '-h' switch which shows help on using the script.
Bugfix: add "-noappend" to the squashfs command in the script to avoid
adding new data to an existing squashfs module.