Disable the SSH daemon by default and use a tweak to start it on demand.
The passwords for users 'live' and 'root' are easy to guess.
When you use the Slackware Live Edition on a public network,
you do not want people to use an exposed SSH login.
If you still want to have the SSH daemon enabled on boot, use the
'tweak' boot parameter and provide a 'ssh' value as follows:
  "tweak=ssh"
This can be combined with other tweaks of course, all comma-separated.