liveinit: use boot parameter "nop=wipe" to erase existing persistent data.
Sometimes your Slackware Live develops issues because the persistent data
storage is conflicting with new packages or data which you added to the
Live OS.  This can be fixed by erasing persistent data, thereby reverting
the Live OS to its initial state - i.e. when the USB stick was booted
for the first time.
For this purpose the functionality of the existing "nop" boot parameter
was extended to allow the '=' and specifying a value of 'wipe'.

Note that the persistent data will be wiped before the Live OS boots, and
it does not matter if you are using a persistence directory or a container
file for storing persistent data.

This wipe includes all your persistent personal data in /home/
unless you have moved your /home into a container file on the USB stick.