Add '-G' parameter to generate ISO image from pre-existing file tree.
This is of course a new parameter to the "make_slackware_live.sh" script.
It was added to allow the developer to make small changes to the file tree
of what is to become the ISO image, and then re-generate the ISO without
having to repeat all the other steps (potentially undoing the modifications).
Useful for debugging mostly.