Make the name 'min' of the minimal package list customizable.
The 'min' package list contains the generic kernel.
This package list is special because the 'make_slackware_live.sh' script
will also take care of the ISO boot configuration while processing it.
A custom Live OS may want to use a custom filename instead of "min.lst",
so the variable $MINLIST was created for that purpose.