Make it possible to replace a package instead of simply upgrading it.
Example:
Replace eudev with systemd (actual request made by Dlackware).
The package list should then contain the following line:

eudev%systemd

I.e. the package to be replaced, followed by a '%', followed by the
name of the package that replaces the original.
This is similar syntax to the way in which you tell upgradepkg to
replace one package with another bearing a different name.