The liveinit script is a template too and cannot be used as-is.
This file was forgotten in the previous commit.
It has now also been given a '.tpl' extension in the liveslak sources,
to indicate that it is a template and not usable as-is.
The 'make_slackware_live.sh' script parses the template file and performs
a series of variable substitutions to generate the final 'init' script
inside the initrd of the ISO image.