Add boot-time tweak 'nsh' to disable new sub-pixel hinting in freetype
Freetype has enabled a 'new style' of sub-pixel hinting, based on the
Infinality patch but simplified for speed.
Some people feel that this new type of hinting is resulting in worse font
rendering than with the old auto-hinter built into Freetype.
The new tweak 'nsh' effectively disables the new type of sub-pixel hinting
and makes Freetype fall back to the old behaviour.

Usage: add "tweaks=nsh" to the boot commandline of the Live OS.